Reduce as_shield_badges assumptions about git environment
This commit modifies `count_lines_mod::as_shield_badges` to return an empty
string if running in a git repository without any remotes, in a directory
which is not a git repository, or on a system which does not have a git
command available. Previously, the function panicked in all of these cases.